requirements.txt
absl-py==2.1.0
accelerate==0.12.0
contourpy==1.3.2
cycler==0.12.1
einops==0.8.1
filelock==3.13.1
flash-attn==2.7.3
fonttools==4.57.0
fsspec==2024.6.1
grpcio==1.71.0
importlib_resources==6.5.2
Jinja2==3.1.4
joblib==1.5.0
kiwisolver==1.4.8
Markdown==3.8
MarkupSafe==2.1.5
matplotlib==3.10.1
ml_collections==1.0.0
mpmath==1.3.0
networkx==3.3
nibabel==5.3.2
ninja==1.11.1.4
numpy==2.1.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
packaging==25.0
pandas==2.2.3
pillow==11.0.0
protobuf==6.30.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
timm==0.3.2
torch==2.6.0+cu124
torchvision==0.21.0+cu124
tqdm==4.67.1
triton==3.2.0
typing_extensions==4.12.2
tzdata==2025.2
Werkzeug==3.1.3
